+++ bank-feeds/RELEASES.md
@@ -717,3 +717,13 @@ Based on:
 - [typescript v2.4.1] bank-feeds
 ### Releases
 - [NPM v2.4.1] https://www.npmjs.com/package/@codat/bank-feeds/v/2.4.1 - bank-feeds
+
+## 2023-10-03 00:05:29
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.93.1 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.4.2] bank-feeds
+### Releases
+- [NPM v2.4.2] https://www.npmjs.com/package/@codat/bank-feeds/v/2.4.2 - bank-feeds